refactor cli main file
